# packages/connection-contracts/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Connection Browser API",
		"version": "1.0.0",
		"description": "Connection Web 使用的同源 HTTP/JSON 契约。当前身份只由 HttpOnly browser session 解析。"
	},
	"servers": [{ "url": "/" }],
	"security": [{ "browserSession": [] }],
	"paths": {
...
		"/api/v1/connection/connections/{connectionId}": {
			"delete": {
				"operationId": "disconnectConnection",
				"parameters": [
					{ "$ref": "#/components/parameters/ConnectionId" },
					{ "$ref": "#/components/parameters/IdempotencyKey" }
				],
				"responses": {
					"204": { "description": "Disconnected" },
					"401": { "$ref": "#/components/responses/Error" },
					"404": { "$ref": "#/components/responses/Error" },
					"409": { "$ref": "#/components/responses/Error" }
				}
			}
		},
		"/api/v1/connection/oauth-transactions": {
			"post": {
				"operationId": "startGithubOAuth",
				"parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OAuthTransactionRequest"
							}
						}
					}
				},
				"responses": {
					"200": { "$ref": "#/components/responses/OAuthTransaction" },
					"401": { "$ref": "#/components/responses/Error" },
					"404": { "$ref": "#/components/responses/Error" },
					"409": { "$ref": "#/components/responses/Error" }
				}
			}
		},
...
		"/api/v1/connection/grants/{grantId}": {
			"delete": {
				"operationId": "revokeGrant",
				"parameters": [
					{ "$ref": "#/components/parameters/GrantId" },
					{ "$ref": "#/components/parameters/IdempotencyKey" }
				],
				"responses": {
					"204": { "description": "Revoked" },
					"401": { "$ref": "#/components/responses/Error" },
					"404": { "$ref": "#/components/responses/Error" },
					"409": { "$ref": "#/components/responses/Error" }
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"browserSession": {
				"type": "apiKey",
				"in": "cookie",
				"name": "connection_session"
			}
		},
		"parameters": {
			"IdempotencyKey": {
				"name": "Idempotency-Key",
				"in": "header",
				"required": true,
				"schema": {
					"type": "string",
					"minLength": 8,
					"maxLength": 200,
					"pattern": "^[\\x21-\\x7E]+$"
				}
			},
			"ConnectionId": {
				"name": "connectionId",
				"in": "path",
				"required": true,
				"schema": { "type": "string", "minLength": 1 }
			},
			"GrantId": {
				"name": "grantId",
				"in": "path",
				"required": true,
				"schema": { "type": "string", "minLength": 1 }
			},
...
		},
		"responses": {
			"Error": {
				"description": "Stable browser error",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" }
					}
				}
			},
...
			"OAuthTransaction": {
				"description": "Authorization URL",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/OAuthTransaction" }
					}
				}
			},
...
		},
		"schemas": {
			"Error": {
				"type": "object",
				"additionalProperties": false,
				"required": ["error"],
				"properties": {
					"error": {
						"type": "object",
						"additionalProperties": false,
						"required": ["code", "messageKey", "retryable", "traceId"],
						"properties": {
							"code": { "type": "string" },
							"messageKey": { "type": "string" },
							"retryable": { "type": "boolean" },
							"traceId": { "type": "string" }
						}
					}
				}
			},
...
			"OAuthTransactionRequest": {
				"type": "object",
				"additionalProperties": false,
				"properties": { "sharedScopeId": { "type": "string" } }
			},
			"OAuthTransaction": {
				"type": "object",
				"additionalProperties": false,
				"required": ["authorizationUrl"],
				"properties": {
					"authorizationUrl": { "type": "string", "format": "uri" }
				}
			},
...
		}
	}
}